# docs/v1/v1_swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdminChangeRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  dto.AdminInfoResponse:
    properties:
      avatar:
        type: string
      id:
        type: integer
      introduction:
        type: string
      login_time:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  dto.AdminLoginRequest:
    properties:
      password:
        example: admin
        type: string
      username:
        example: admin
        type: string
    type: object
  dto.AdminLoginResponse:
    properties:
      token:
        type: string
    type: object
...
  middleware.ResponseErr:
    properties:
      data: {}
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: string
    type: object
host: 127.0.0.1:2401
info:
  contact:
    email: stone2401@qq.com
  description: 一个轻量级网关代理，支持多种代理协议，接口转发，数据统计，带有管理界面
  title: 轻量级网关
  version: "1.0"
paths:
  /admin/admin_info:
    get:
      consumes:
      - application/json
      description: 获取admin信息
      operationId: /admin/admin_info
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminInfoResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: admin基本信息
      tags:
      - adminApi
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: 修改密码
      operationId: /admin/change_pwd
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: admin密码修改
      tags:
      - adminApi
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 描述
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminLoginResponse'
              type: object
      summary: admin登录接口
      tags:
      - adminApi
...
schemes:
- http
swagger: "2.0"